Fix broken S-band model
As reported in JIRA SPR1-1789 this was raising an exception when trying to load
S-band models (thus breaking the online imaging for S-band). The issue was the
extra carriage return at the end of the docstring defining the beams, this was
being read as an empty line and causing `numpy.loadtxt` some grief.